refactor: share threads list between agent / networks and cloud
