+++ packages/cli/src/playground/package.json
@@ -16,7 +16,6 @@
     "@lezer/highlight": "^1.2.1",
     "@mastra/client-js": "workspace:*",
     "@mastra/playground-ui": "workspace:^",
-    "@radix-ui/react-alert-dialog": "^1.1.14",
     "@radix-ui/react-checkbox": "^1.3.2",
     "@radix-ui/react-collapsible": "^1.1.11",
     "@radix-ui/react-dialog": "^1.1.14",
