+++ packages/playground-ui/src/domains/agents/index.tsx
@@ -7,3 +7,4 @@ export * from './components/runtime-context';
 export * from './components/provider-map-icon';
 export * from './components/agent-metadata';
 export * from './components/agent-entity-header';
+export * from './components/chat-threads';
