+++ packages/playground-ui/src/index.ts
@@ -36,3 +36,4 @@ export type { TraceContextType } from './domains/traces/context/trace-context';
 export * from './store/playground-store';
 export * from './lib/framework';
 export { MemorySearch } from './components/assistant-ui/memory-search';
+export * from './components/ui/alert-dialog';
